+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
--
+- feature: Web Push notifications frontend (Service Worker + Push subscription flow + settings UI + open conversation on notification click)
 
 # [Release]